Changed Stock Quote App to Use XML Pull Parser
Parsed XML using XML pull parser instead of using the dom because this
is the method that Google recommends. The app crashes because of
ArrayIndexOutOfBoundsException. I can't seem to figure out why.  Also I
dont understand the beginDocument method and how it is working as well
as the do block in the doInBackGroundMethod.